requirements.txt
streamlit
openai==0.28
python-dotenv
pytesseract
Pillow
